+++ src/http/DurationAdapter.java
@@ -1,4 +1,4 @@
-package httpServer;
+package http;
 
 import com.google.gson.TypeAdapter;
 import com.google.gson.stream.JsonReader;
