+++ src/http/handler/BaseHttpHandler.java
@@ -1,4 +1,4 @@
-package httpServer.handler;
+package http.handler;
 
 import com.sun.net.httpserver.HttpExchange;
