+++ src/http/handler/PrioritizedHandler.java
@@ -1,8 +1,8 @@
-package httpServer.handler;
+package http.handler;
 
 import com.sun.net.httpserver.HttpExchange;
 import com.sun.net.httpserver.HttpHandler;
-import httpServer.HttpTaskServer;
+import http.HttpTaskServer;
 import tasks.Task;
 
 import java.io.IOException;
