+++ test/http/HttpServerHistoryTest.java
@@ -1,4 +1,4 @@
-package httpServer;
+package http;
 
 import com.google.gson.Gson;
 import io.restassured.response.Response;
